+++ src/thumbor_video_engine/engines/video.py
@@ -1,10 +1,7 @@
-from io import BytesIO
-
-from PIL import Image
 from thumbor.engines import BaseEngine
 from thumbor.utils import logger
 
-from thumbor_video_engine.utils import named_tmp_file, is_mp4
+from thumbor_video_engine.utils import named_tmp_file, is_mp4, is_animated, is_animated_gif
 
 
 def patch_baseengine_get_mimetype():
@@ -24,11 +21,6 @@ def patch_baseengine_get_mimetype():
 
 
 patch_baseengine_get_mimetype()
-
-
-def is_animated(buffer):
-    im = Image.open(BytesIO(buffer))
-    return getattr(im, 'is_animated', False)
 
 
 class Engine(object):
@@ -61,33 +53,51 @@ class Engine(object):
                 self.context.modules.importer.ffmpeg_engine(self.context))
         return self.context.modules.ffmpeg_engine
 
-    def load(self, buffer, extension):
+    def get_engine(self, buffer, extension):
         mime = BaseEngine.get_mimetype(buffer)
 
         is_gif = extension == '.gif'
         is_webp = extension == '.webp'
+        accepts_video = getattr(self.context.request, "accepts_video", False)
+        accepts_webp = self.context.request.accepts_webp
 
         if is_webp and self.ffmpeg_handle_animated_webp and is_animated(buffer):
-            logger.debug("Setting engine to %s (extension %s)" % (
-                self.context.config.FFMPEG_ENGINE, extension))
-            self.engine = self.ffmpeg_engine
-        elif is_gif and self.ffmpeg_handle_animated_gif and is_animated(buffer):
-            logger.debug("Setting engine to %s (extension %s)" % (
-                self.context.config.FFMPEG_ENGINE, extension))
-            self.engine = self.ffmpeg_engine
+            return self.ffmpeg_engine
+        elif is_gif and self.ffmpeg_handle_animated_gif and is_animated_gif(buffer):
+            if self.context.config.FFMPEG_GIF_AUTO_H265:
+                self.context.request.should_vary = True
+                if accepts_video:
+                    logger.debug("FFMPEG_GIF_AUTO_H265 setting format to h264")
+                    self.context.request.format = 'h265'
+            elif self.context.config.FFMPEG_GIF_AUTO_H264:
+                self.context.request.should_vary = True
+                if accepts_video:
+                    logger.debug("FFMPEG_GIF_AUTO_H264 setting format to h264")
+                    self.context.request.format = 'h264'
+            elif self.context.config.FFMPEG_GIF_AUTO_WEBP:
+                self.context.request.should_vary = True
+                if accepts_webp:
+                    logger.debug("FFMPEG_GIF_AUTO_WEBP setting format to webp")
+                    self.context.request.format = 'webp'
+            return self.ffmpeg_engine
         elif is_gif and self.use_gif_engine:
-            logger.debug("Setting engine to %s (extension %s)" % (
-                self.context.config.GIF_ENGINE, extension))
-            self.engine = self.context.modules.gif_engine
+            return self.context.modules.gif_engine
         elif mime.startswith('video/'):
-            logger.debug("Setting engine to %s (extension %s)" % (
-                self.context.config.FFMPEG_ENGINE, extension))
-            self.engine = self.ffmpeg_engine
+            return self.ffmpeg_engine
         else:
-            logger.debug("Setting engine to %s (extension %s)" % (
-                self.context.config.IMAGE_ENGINE, extension))
-            self.engine = self.image_engine
+            return self.image_engine
 
+    def load(self, buffer, extension):
+        self.engine = self.get_engine(buffer, extension)
+        if self.context.request.format and not self.context.request.filters:
+            # RequestParameters.filters is an empty list when none are in the url,
+            # and ImagingHandler._write_results_to_client assumes that if
+            # context.request.format is set then it came from the format filter.
+            # Since we set the format in the engine this causes a TypeError,
+            # so we need to ensure that it is a string here.
+            self.context.request.filters = ""
+        logger.debug("Set engine to %s (extension %s)" % (
+            type(self.engine).__module__, extension))
         still_frame_pos = getattr(self.context.request, 'still_position', None)
         # Are we requesting a still frame?
         if self.engine is self.ffmpeg_engine and still_frame_pos: