+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
 
-from thumbor_video_engine.utils import is_mp4
+from thumbor_video_engine.utils import is_mp4, is_animated_gif
 
 
 @pytest.mark.parametrize('bool_val,buf', [
@@ -15,3 +15,18 @@ from thumbor_video_engine.utils import is_mp4
 ])
 def test_is_mp4(bool_val, buf):
     assert is_mp4(buf) is bool_val
+
+
+@pytest.mark.parametrize('buf', [
+    b'',
+    b'GIF89a\xc8\x00\x96\x00\x00\x00\x00\x00',
+    b'GIF89a\x01\x00\x01\x00\x00\xff\x00,\x00\x00\x00\x00\x01\x00\x01\x00\x00\x02\x00;',
+])
+def test_is_animated_gif_false(buf):
+    assert is_animated_gif(buf) is False
+
+
+def test_is_animated_gif_true(storage_path):
+    with open("%s/hotdog.gif" % storage_path, mode="rb") as f:
+        im_bytes = f.read()
+    assert is_animated_gif(im_bytes) is True